fix(modal): Split component
