+++ modules/modal/react/package.json
@@ -39,7 +39,7 @@
     "@emotion/styled": "^10.0.17",
     "@workday/canvas-kit-react-common": "^3.1.0",
     "@workday/canvas-kit-react-popup": "^3.1.0",
-    "focus-trap-js": "1.0.7"
+    "focus-trap-js": "1.0.8"
   },
   "peerDependencies": {
     "react": ">= 16.8 < 17"
